Move From<WalkdirError> for HolochainError into cas_implementations
