+++ core_types/Cargo.toml
@@ -10,7 +10,6 @@ serde_json="1"
 multihash = "0.8.0"
 futures-preview = "0.2.2"
 riker = "0.1.7"
-walkdir = "2"
 rust-base58 = "0.0.4"
 snowflake = "1.2"
 unwrap_to = "0.1.0"
