+++ core_types/src/error.rs
@@ -7,9 +7,7 @@ use std::{
     error::Error,
     fmt,
     io::{self, Error as IoError},
-    path::Path,
 };
-use walkdir::Error as WalkdirError;
 
 /// Enum holding all Holochain specific errors
 #[derive(Clone, Debug, PartialEq, Hash, Eq)]
@@ -75,18 +73,6 @@ fn reason_for_io_error(error: &IoError) -> String {
         io::ErrorKind::InvalidData => format!("contains invalid data: {}", error),
         io::ErrorKind::PermissionDenied => format!("missing permissions to read: {}", error),
         _ => format!("unexpected error: {}", error),
-    }
-}
-
-impl From<WalkdirError> for HolochainError {
-    fn from(error: WalkdirError) -> Self {
-        // adapted from https://docs.rs/walkdir/2.2.5/walkdir/struct.Error.html#example
-        let path = error.path().unwrap_or(Path::new("")).display();
-        let reason = match error.io_error() {
-            Some(inner) => reason_for_io_error(inner),
-            None => String::new(),
-        };
-        HolochainError::IoError(format!("error at path: {}, reason: {}", path, reason))
     }
 }
